updates on the soap microservice

+++ matieres/src/main/java/com/nabou/matieres/repositories/MatiereRepository.java
@@ -3,5 +3,5 @@ package com.nabou.matieres.repositories;
 import com.nabou.matieres.entities.Matiere;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MatiereRepository extends JpaRepository<Matiere, Long> {
+public interface MatiereRepository extends JpaRepository<Matiere, Integer> {
 }
